+++ src/main/java/com/imb2025/calificaciones/service/jpa/MateriaServiceImpl.java
@@ -9,60 +9,58 @@ import com.imb2025.calificaciones.entity.Materia;
 import com.imb2025.calificaciones.repository.MateriaRepository;
 import com.imb2025.calificaciones.service.IMateriaService;
 
-
 @Service
 public class MateriaServiceImpl implements IMateriaService{
-	
-	@Autowired
-	private MateriaRepository repo;
 
-	@Override
-	public List<Materia> findAll() {
-		
-		return repo.findAll();
-		
-	}
+    @Autowired
+    private MateriaRepository repo;
 
-	@Override
-	public Materia findById(Long id) {
-	
-		
-		
-		return repo.findById(id).orElse(null);
-	}
+    @Override
+    public List<Materia> findAll() {
 
-	@Override
-	public Materia create (Materia materia) {
-		
-		return repo.save(materia);
-	}
-	
-	@Override
-	public Materia update(Materia materia, Long id) throws Exception {
-		if(repo.existsById(id)) {
-			materia.setId(id);
-			return repo.save(materia);
-		}else {
-			throw new Exception("No se encontró materia con id" + id);
-		}
-		
-	}
-	
-	
-	@Override
-	public void deleteById(long id) {
-		repo.deleteById(id);
-		
-	}
+        return repo.findAll();
 
-	@Override
-	public Materia mapFromDto(MateriaRequestDto materiaRequestDto) throws Exception {
-		Materia materia = new Materia();
-		materia.setNombre(materiaRequestDto.getNombre());
-		materia.setCargaHoraria(materiaRequestDto.getCargaHoraria());
-		materia.setCodigo(materiaRequestDto.getCodigo());
-		materia.setNivel(materiaRequestDto.getNivel());
-		
-		return materia;
-	}
+    }
+
+    @Override
+    public Materia findById(Long id) {
+
+        return repo.findById(id).orElse(null);
+    }
+
+    @Override
+    public Materia create (Materia materia) {
+
+        return repo.save(materia);
+    }
+
+    @Override
+    public Materia update(Materia materia, Long id) throws Exception {
+        if(repo.existsById(id)) {
+            materia.setId(id);
+            return repo.save(materia);
+        }else {
+            throw new Exception("No se encontró materia con id" + id);
+        }
+
+    }
+
+    @Override
+    public void deleteById(Long id) throws Exception {
+        if (!repo.existsById(id)) {
+            throw new Exception("No se puede eliminar el id: " + id + " porque no existe");
+        }
+        repo.deleteById(id);
+    }
+
+    @Override
+    public Materia fromDto(MateriaRequestDto materiaRequestDto) throws Exception {
+        Materia materia = new Materia();
+        materia.setNombre(materiaRequestDto.getNombre());
+        materia.setCargaHoraria(materiaRequestDto.getCargaHoraria());
+        materia.setCodigo(materiaRequestDto.getCodigo());
+        materia.setNivel(materiaRequestDto.getNivel());
+
+        return materia;
+    }
 }
